Add support for retreiving PostScript files
This commit adds support for retreiving the generated PostScript .ps
file from AVL as the default behavior as opposed having to install
GhostScript to convert this file. Both Adobe Illustrator and InkScape
support PostScript, and using these tools to view the plots is an
alternative usage for the save_geometry_plot and save_trefftz_plots
methods. Without this fix, the user would be forced to install
GhostScript to get any plot output from AVL, which violates having
GhostScript as an optional dependency as per the README.

Summary of Changes:
- Correct README to specify that GhostScript is only required to
  convert and save plots to pdf or raster formats.
- Add condition to _get_plot to copy the output .ps file to the
  target_dir and return without continuing onto conversion using
  GhostScript
- Modify default value of the file_format argument to .ps (Potentially
  a breaking change for some, but this fix is inline with how
  the README treats GhostScript as an optional dependency)